web-api/requirements.txt
aiodns==3.2.0
aiohttp==3.10.10
asyncio==3.4.3
azure-identity==1.19.0
boto3>=1.34.0
fastapi==0.115.2
hashids==1.3.1
openai==1.59.7
pydantic==2.9.2
pydantic-settings==2.6.0
pydub==0.25.1
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.12,
sqids==0.5.0
SQLAlchemy==2.0.36
uvicorn==0.32.0
numpy>=2.2.5
mlx>=0.0.7
parakeet-mlx>=0.2.7
httpx>=0.27.0